+++ src/test/helper/AppClassWithInvalidBindingModule.php
@@ -7,8 +7,8 @@
  *
  * @package  stubbles
  */
-namespace stubbles\test\ioc;
-use stubbles\ioc\App;
+namespace stubbles\test;
+use stubbles\App;
 /**
  * Helper class for ioc tests.
  */